#151: Sorted out the source files
