+++ doc/changes/unreleased.md
@@ -1,1 +1,5 @@
 # Unreleased
+
+## Refactorings
+
+* #151: Sorted out the source files and tests into directories.
