+++ exasol/ai/mcp/server/setup/generic_auth.py
