+++ exasol/ai/mcp/server/setup/server_settings.py
