+++ exasol/ai/mcp/server/tools/data/exasol_builtin_functions.json
