+++ exasol/ai/mcp/server/tools/parameter_pattern.py
