+++ exasol/ai/mcp/server/utils/keyword_search.py
