+++ exasol/ai/mcp/server/utils/named_object_pool.py
