+++ test/utils/tool_utils.py
@@ -1,0 +1,30 @@
+import asyncio
+from collections.abc import Generator
+from contextlib import contextmanager
+
+from fastmcp import Client
+from pyexasol import ExaConnection
+
+from exasol.ai.mcp.server.connection.db_connection import DbConnection
+from exasol.ai.mcp.server.main import create_mcp_server
+from exasol.ai.mcp.server.setup.server_settings import McpServerSettings
+
+
+async def _run_tool_async(
+    connection: ExaConnection, config: McpServerSettings, tool_name: str, **kwargs
+):
+    @contextmanager
+    def connection_factory() -> Generator[ExaConnection, None, None]:
+        yield connection
+
+    db_connection = DbConnection(connection_factory, num_retries=1)
+
+    exa_server = create_mcp_server(db_connection, config)
+    async with Client(exa_server) as client:
+        return await client.call_tool(tool_name, kwargs)
+
+
+def run_tool(
+    connection: ExaConnection, config: McpServerSettings, tool_name: str, **kwargs
+):
+    return asyncio.run(_run_tool_async(connection, config, tool_name, **kwargs))
